split out documentation into seperate directory
fix test/examples.js require

fix requires

temp

symlink test

update description
