+++ .github/dependabot.yml
@@ -12,6 +12,12 @@ updates:
       interval: "weekly"
     open-pull-requests-limit: 10
 
+  - package-ecosystem: "npm"
+    directory: /documentation
+    schedule:
+      interval: "weekly"
+    open-pull-requests-limit: 10
+
   - package-ecosystem: docker
     directory: /build
     schedule:
