+++ .npmignore
@@ -9,4 +9,4 @@ lib/llhttp/llhttp.wasm
 
 !types/**/*
 !index.d.ts
-!docs/**/*
+!documentation/docs/**/*
