+++ documentation/examples/proxy/index.js
@@ -1,4 +1,4 @@
-const { Pool, Client } = require('../../')
+const { Pool, Client } = require('../../../')
 const http = require('node:http')
 const proxy = require('./proxy')
